Query now has variables
A query can now have named sprintf style placeholders (i.e.
%{start_date} ) embeded in them. In order to make this work, any %
%(percent) signs must be changed to %% (double percent sign).

Private helper methods were added to the query model to find all named
placeholders in the command and then add them as keys in the variables
hash (value set to nil)

Exceptions will be raised if a command requires a variable that is not
set. We will need to decide how to deal with this later.