Switch to patched version of net-ldap gem
We need version 0.3.1 of the net-ldap gem for devise. However, we also
need this patch
(https://github.com/ruby-ldap/ruby-net-ldap/commit/401f3a3c33b7f1c82b6074a5f63d4cbf0e385fca)
which is only available in a newer version of the gem on github. To get
around this, I have forked version 0.3.1 of the gem, applied the patch,
and uploaded it to github.